third_party/blink/web_tests/platform/win/external/wpt/css/css-fonts/cjk-kerning-expected.txt

This is a testharness.js-based test.
[FAIL] expected match: .kernON .cjk vs .paltONkernON .cjk
  assert_equals: expected 308 but got 380
[FAIL] expected match: .default .cjk vs .kernOFF .cjk
  assert_equals: expected 384 but got 380
Harness: the test ran to completion.
